Added local connection to startup ui
